+++ embeddings.js
@@ -2,6 +2,8 @@ import fs from "fs/promises";
 import { mkdirp } from "mkdirp";
 import { glob } from "glob";
 import ollama from "ollama";
+import { default as cliProgress } from "cli-progress";
+import { existsSync } from "fs";
 
 // Constants
 const path = "articles/3-embeddings";
@@ -17,7 +19,6 @@ async function embedding(article) {
                 prompt: chunk.content,
                 model,
             })
-            // console.log(response)
             chunk.embedding = response.embedding;
         } catch (error) {
             console.log(`Error parsing response: ${response}`);
@@ -27,23 +28,30 @@ async function embedding(article) {
     return chunks;
 }
 
-async function main() {
+export async function Embeddings() {
     const time_start = Date.now();
     await mkdirp(path);
     const article_files = await glob(`${previous_path}/*.json`);
-    console.log(`Found ${article_files.length} articles.`);
+    const bar1 = new cliProgress.SingleBar({}, cliProgress.Presets.shades_classic);
+    bar1.start(article_files.length, 0);
     let i = 0;
     for (let article_file of article_files) {
         const article = JSON.parse(await fs.readFile(article_file, "utf8"));
+        if (existsSync(`${path}/${article._id}.json`)) {
+            bar1.increment();
+            continue;
+        }
         article.chunks = await embedding(article);
         await fs.writeFile(`${path}/${article._id}.json`, JSON.stringify(article, null, 2));
         i++;
+        bar1.increment();
     }
+    bar1.stop();
     const time_end = Date.now();
     console.log(`Time elapsed: ${(time_end - time_start) / 1000}s`);
     console.log(`Average time per article: ${(time_end - time_start) / i / 1000}s`);
     return true;
 }
 
-main()
-    .catch(console.error)
+// Embeddings()
+//     .catch(console.error)
